Split up libpreflight in multiple modules and improve the storage code
